State Overhaul, API Wrapper skeleton, and Settings Consolidation
I've reworked the Vuex Store to handle the fetching of data so it can be used across the app.
I've also started a jsonbin.org api wrapper to ease the use of the state mechanism.
I've consolidated the defined variables into the modules.exports so it is easy to adjust settings in one place.